src/learnable_gated_pooling/requirements.txt
torch>=2.0.0
numpy>=1.21.0
pandas>=1.3.0
pyyaml>=5.4.0
scikit-learn>=1.0.0
tqdm>=4.65.0
matplotlib>=3.4.0
